+++ R/mod-results-boxes.R
@@ -48,7 +48,9 @@ results_boxes_ui <- function(id) {
       uiOutput(ns("successes")),
       solidHeader = TRUE,
       collapsible = TRUE,
+      collapsed = TRUE,
       title = textOutput(ns("num_success")),
+      footer = "Your files passed these checks. Well done!",
       status = "success",
       width = 12
     ),
@@ -57,6 +59,7 @@ results_boxes_ui <- function(id) {
       solidHeader = TRUE,
       collapsible = TRUE,
       title = textOutput(ns("num_warn")),
+      footer = "Warnings highlight possible issues in the data, but they may not be applicable to every dataset. If they are not applicable for your data, you can ignore them. For example, if any columns are completely empty they will generate a warning. If you did not collect data for that column, then you can ignore the warning.", # nolint
       status = "warning",
       width = 12
     ),
@@ -65,6 +68,7 @@ results_boxes_ui <- function(id) {
       solidHeader = TRUE,
       collapsible = TRUE,
       title = textOutput(ns("num_fail")),
+      footer = "Any failures in this box must be corrected. A summary of data uploaded can be found in the Data Summary tab, and may help discover the reason behind the failures. Contact the DCC team if you have any questions.", # nolint
       status = "danger",
       width = 12
     )